refactor: DivWIthLabel -> DivWithLabel
